requirements.txt
flask
flask-sqlalchemy
flask-cors
python-dotenv
mysql-connector-python
razorpay
PyJWT
Werkzeug
flask-limiter
pytz
